Optimización de código con getCurrentUser

+++ src/main/java/com/reboluxBurger/backend/service/CategoryService.java
@@ -1,7 +1,10 @@
 package com.reboluxBurger.backend.service;
 
 import com.reboluxBurger.backend.entity.Category;
+import com.reboluxBurger.backend.entity.User;
+import com.reboluxBurger.backend.enums.Role;
 import com.reboluxBurger.backend.repository.CategoryRepository;
+import com.reboluxBurger.backend.security.CurrentUserProvider;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -11,9 +14,11 @@ import java.util.Optional;
 public class CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final CurrentUserProvider currentUserProvider;
 
-    public CategoryService(CategoryRepository categoryRepository) {
+    public CategoryService(CategoryRepository categoryRepository, CurrentUserProvider currentUserProvider) {
         this.categoryRepository = categoryRepository;
+        this.currentUserProvider = currentUserProvider;
     }
 
     // Devuelve todas las categorías
@@ -23,18 +28,30 @@ public class CategoryService {
 
     // Crea una nueva categoría si no existe otra con ese nombre
     public Category createCategory(Category category) {
-        if (categoryRepository.existsByName(category.getName())) {
-            throw new RuntimeException("Ya existe una categoría con ese nombre");
+
+        User currentUser = currentUserProvider.getCurrentUser();
+        if (currentUser.getRole() == Role.ADMIN) {
+            if (categoryRepository.existsByName(category.getName())) {
+                throw new RuntimeException("Ya existe una categoría con ese nombre");
+            }
+            return categoryRepository.save(category);
+        } else {
+            throw new RuntimeException("No tienes permisos para crear una categoría");
         }
-        return categoryRepository.save(category);
     }
 
     // Borra una categoría por id si existe
     public void deleteCategory(Long id) {
-        if (!categoryRepository.existsById(id)) {
-            throw new RuntimeException("Categoría no encontrada");
-        }
-        categoryRepository.deleteById(id);
+
+        User currentUser = currentUserProvider.getCurrentUser();
+        if (currentUser.getRole() == Role.ADMIN) {
+            if (!categoryRepository.existsById(id)) {
+                throw new RuntimeException("Categoría no encontrada");
+            }
+            categoryRepository.deleteById(id);
+            } else {
+                throw new RuntimeException("No tienes permisos para eliminar una categoría");
+            }
     }
 
     // Obtiene una categoría por nombre, o la crea si no existe
@@ -47,6 +64,4 @@ public class CategoryService {
             return categoryRepository.save(new Category(null, normalized, imageUrl));
         }
     }
-
-
 }
